Move price stat to his own component that is updated every 60s
