+++ app/Http/Livewire/PriceTicker.php
@@ -16,7 +16,6 @@ final class PriceTicker extends Component
     /** @phpstan-ignore-next-line */
     protected $listeners = [
         'currencyChanged' => 'setValues',
-        'updatePrice'     => 'setValues',
     ];
 
     public string $price;
