+++ resources/views/livewire/price-ticker.blade.php
@@ -1,4 +1,5 @@
 <div
+    wire:poll.60s
     class="uppercase"
     :class="{ 'opacity-50': busy }"
     x-data="{ to: '{{ $to }}', busy: false }"
